refactor(pipeline): own run activity accounting
