+++ docs/topology-status.md
@@ -29,9 +29,9 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 581
+- **Stable** (no move scoped): 582
 - **Kernel** (polylogue/ root): 10
-- **Primitives** (lib-root + storage-root): 18
+- **Primitives** (lib-root + storage-root): 17
 - **TBD** (cell needs explicit assignment): 0
 - **Total declared**: 696
 - **Realized polylogue/**/*.py**: 696 files declared
